[ENH] factor out common _evaluate for classification and regression, fix contributors
- Create sktime/utils/_panel_evaluate.py with shared evaluate logic
- Refactor classification evaluate to thin wrapper calling shared code
- Refactor regression evaluate to thin wrapper calling shared code
- Fix __author__ tags from jgyasu to NAME-ASHWANIYADAV in regression files
- Revert manual CONTRIBUTORS.md edit, update .all-contributorsrc instead
