+++ .all-contributorsrc
@@ -2868,6 +2868,15 @@
         "code",
         "doc"
       ]
+    },
+    {
+      "login": "NAME-ASHWANIYADAV",
+      "name": "NAME-ASHWANIYADAV",
+      "avatar_url": "https://avatars.githubusercontent.com/NAME-ASHWANIYADAV?v=4",
+      "profile": "https://github.com/NAME-ASHWANIYADAV",
+      "contributions": [
+        "code"
+      ]
     }
   ]
 }
