+++ sktime/classification/model_evaluation/_functions.py
@@ -7,36 +7,23 @@ __all__ = ["evaluate"]
 
 import collections.abc
 import inspect
-import time
-import warnings
 
 import numpy as np
-import pandas as pd
-from skbase.utils.dependencies import _check_soft_dependencies
 
-from sktime.datatypes import check_is_scitype, convert
-from sktime.exceptions import FitFailedWarning
-from sktime.utils.parallel import parallelize
-
-PANDAS_MTYPES = ["pd.DataFrame", "pd.Series", "pd-multiindex", "pd_multiindex_hier"]
+from sktime.utils._panel_evaluate import _run_evaluate
 
 
-def _is_proba_classification_score(metric) -> bool:
-    """
-    Check if metric function is intended for probabilistic classification.
+def _is_proba_classification_score(metric):
+    """Determine whether a metric requires predict_proba output.
 
-    This function attempts to identify if the input `metric` is a classification
-    score that expects predicted probabilities rather than class labels.
-    It performs this check using:
-    1. A set of known probability-based metric names.
-    2. Inspection of the function's signature for indicative argument names.
-    3. A test call using sample `y_true` and `y_pred_proba` arrays to check
-    compatibility.
+    Uses heuristics to check whether a given sklearn-compatible metric function
+    is for probabilistic classification (i.e., expects probability predictions
+    rather than discrete class labels).
 
     Parameters
     ----------
-    metric : callable or str
-    A metric function or its name to check.
+    metric : callable
+        A scoring function following the sklearn metric interface.
 
     Returns
     -------
@@ -109,155 +96,6 @@ def _check_scores(metrics) -> dict:
             metrics_type[scitype].append(metric)
 
     return metrics_type
-
-
-def _get_column_order_and_datatype(
-    metric_types: dict, return_data: bool = True
-) -> dict:
-    """Get the ordered column name and input datatype of results."""
-    y_metadata = {
-        "X_train": "object",
-        "X_test": "object",
-        "y_train": "object",
-        "y_test": "object",
-    }
-    fit_metadata, metrics_metadata = {"fit_time": "float"}, {}
-    for scitype in metric_types:
-        for metric in metric_types.get(scitype):
-            pred_args = _get_pred_args_from_metric(scitype, metric)
-            if pred_args == {}:
-                time_key = f"{scitype}_time"
-                result_key = f"test_{metric.__name__}"
-                y_pred_key = f"y_{scitype}"
-            else:
-                argval = list(pred_args.values())[0]
-                time_key = f"{scitype}_{argval}_time"
-                result_key = f"test_{metric.__name__}_{argval}"
-                y_pred_key = f"y_{scitype}_{argval}"
-            fit_metadata[time_key] = "float"
-            metrics_metadata[result_key] = "float"
-            if return_data:
-                y_metadata[y_pred_key] = "object"
-    if return_data:
-        fit_metadata.update(y_metadata)
-    metrics_metadata.update(fit_metadata)
-    return metrics_metadata.copy()
-
-
-def _get_pred_args_from_metric(scitype, metric):
-    pred_args = {
-        "pred_quantiles": "alpha",
-    }
-    if scitype in pred_args.keys():
-        val = getattr(metric, pred_args[scitype], None)
-        if val is not None:
-            return {pred_args[scitype]: val}
-    return {}
-
-
-def _evaluate_fold(x, meta):
-    i, (y_train, y_test, X_train, X_test) = x
-    classifier = meta["classifier"]
-    scoring = meta["scoring"]
-    return_data = meta["return_data"]
-    error_score = meta["error_score"]
-
-    score = error_score
-    fit_time = np.nan
-    pred_time = np.nan
-    y_pred = pd.NA
-    temp_result = dict()
-    y_preds_cache = dict()
-
-    try:
-        # fit
-        start_fit = time.perf_counter()
-
-        classifier = classifier.clone()
-        classifier.fit(y=y_train, X=X_train)
-
-        fit_time = time.perf_counter() - start_fit
-
-        # predict based on metrics
-        pred_type = {
-            "pred_quantiles": "predict_quantiles",
-            "pred_proba": "predict_proba",
-            "pred": "predict",
-        }
-        # cache prediction from the first scitype and reuse it to
-        # compute other metrics
-        for scitype in scoring:
-            method = getattr(classifier, pred_type[scitype])
-            for metric in scoring.get(scitype):
-                pred_args = _get_pred_args_from_metric(scitype, metric)
-                if pred_args == {}:
-                    time_key = f"{scitype}_time"
-                    result_key = f"test_{metric.__name__}"
-                    y_pred_key = f"y_{scitype}"
-                else:
-                    argval = list(pred_args.values())[0]
-                    time_key = f"{scitype}_{argval}_time"
-                    result_key = f"test_{metric.__name__}_{argval}"
-                    y_pred_key = f"y_{scitype}_{argval}"
-
-                # make prediction
-                if y_pred_key not in y_preds_cache.keys():
-                    start_pred = time.perf_counter()
-                    y_pred = method(X_test, **pred_args)
-                    pred_time = time.perf_counter() - start_pred
-                    temp_result[time_key] = [pred_time]
-                    y_preds_cache[y_pred_key] = [y_pred]
-                else:
-                    y_pred = y_preds_cache[y_pred_key][0]
-
-                if scitype == "pred_proba":
-                    if "pos_label" in inspect.signature(metric).parameters:
-                        pos_label = 1
-                        score = metric(y_test, y_pred[:, 1], pos_label=pos_label)
-                    else:
-                        score = metric(y_test, y_pred)
-                else:
-                    score = metric(y_test, y_pred)
-                temp_result[result_key] = [score]
-
-    except Exception as e:
-        if error_score == "raise":
-            raise e
-        else:  # assign default value when fitting failed
-            for scitype in scoring:
-                temp_result[f"{scitype}_time"] = [pred_time]
-                if return_data:
-                    temp_result[f"y_{scitype}"] = [y_pred]
-                for metric in scoring.get(scitype):
-                    temp_result[f"test_{metric.__name__}"] = [score]
-            warnings.warn(
-                f"""
-                In evaluate, fitting of classifier {type(classifier).__name__} failed,
-                you can set error_score='raise' in evaluate to see
-                the exception message.
-                Fit failed for the {i}-th data split, on training data y_train
-                , and len(y_train)={len(y_train)}.
-                The score will be set to {error_score}.
-                Failed classifier with parameters: {classifier}.
-                """,
-                FitFailedWarning,
-                stacklevel=2,
-            )
-
-    # Storing the remaining evaluate detail
-    temp_result["fit_time"] = [fit_time]
-
-    if return_data:
-        temp_result["X_train"] = [X_train]
-        temp_result["X_test"] = [X_test]
-        temp_result["y_train"] = [y_train]
-        temp_result["y_test"] = [y_test]
-        temp_result.update(y_preds_cache)
-    result = pd.DataFrame(temp_result)
-    column_order = _get_column_order_and_datatype(scoring, return_data)
-    result = result.reindex(columns=column_order.keys())
-
-    return result
 
 
 def evaluate(
@@ -389,94 +227,17 @@ def evaluate(
     >>> cv = KFold(n_splits=3, shuffle=False)
     >>> results = evaluate(classifier=classifier, cv=cv, X=X, y=y)
     """
-    if backend in ["dask", "dask_lazy"]:
-        if not _check_soft_dependencies("dask", severity="none"):
-            raise RuntimeError(
-                "running evaluate with backend='dask' requires the dask package "
-                "installed, but dask is not present in the python environment"
-            )
     scoring = _check_scores(scoring)
 
-    # default handling for cv
-    if isinstance(cv, int):
-        from sklearn.model_selection import KFold
-
-        _cv = KFold(n_splits=cv, shuffle=True)
-    elif cv is None:
-        from sklearn.model_selection import KFold
-
-        _cv = KFold(n_splits=3, shuffle=True)
-    else:
-        _cv = cv
-
-    y_valid, _, y_metadata = check_is_scitype(y, scitype="Table", return_metadata=[])
-    if not y_valid:
-        raise TypeError(f"Expected y dtype Table. Got {type(y)} instead.")
-    y_mtype = y_metadata.get("mtype", None)
-
-    y = convert(y, from_type=y_mtype, to_type="pd_DataFrame_Table")
-
-    X_valid, _, X_metadata = check_is_scitype(X, scitype="Panel", return_metadata=[])
-    if not X_valid:
-        raise TypeError(f"Expected X dtype Panel. Got {type(X)} instead.")
-    X_mtype = X_metadata.get("mtype", None)
-    X = convert(X, from_type=X_mtype, to_type=PANDAS_MTYPES)
-
-    _evaluate_fold_kwargs = {
-        "classifier": classifier,
-        "scoring": scoring,
-        "return_data": return_data,
-        "error_score": error_score,
-    }
-
-    def gen_y_X_train_test(y, X, cv):
-        """Generate joint splits of y, X as per cv.
-
-        Yields
-        ------
-        y_train : i-th train split of y as per cv
-        y_test : i-th test split of y as per cv
-        X_train : i-th train split of y as per cv.
-        X_test : i-th test split of y as per cv.
-        """
-        instance_idx = X.index.get_level_values(0).unique()
-
-        for train_instance_idx, test_instance_idx in cv.split(instance_idx):
-            train_instances = instance_idx[train_instance_idx]
-            test_instances = instance_idx[test_instance_idx]
-
-            X_train = X.loc[X.index.get_level_values(0).isin(train_instances)]
-            X_test = X.loc[X.index.get_level_values(0).isin(test_instances)]
-
-            y_train = y.iloc[train_instance_idx]
-            y_test = y.iloc[test_instance_idx]
-
-            yield y_train, y_test, X_train, X_test
-
-    # generator for y and X splits to iterate over below
-    yx_splits = gen_y_X_train_test(y, X, _cv)
-
-    results = parallelize(
-        fun=_evaluate_fold,
-        iter=enumerate(yx_splits),
-        meta=_evaluate_fold_kwargs,
+    return _run_evaluate(
+        estimator=classifier,
+        estimator_type="classifier",
+        cv=cv,
+        scoring=scoring,
+        X=X,
+        y=y,
+        return_data=return_data,
+        error_score=error_score,
         backend=backend,
         backend_params=backend_params,
     )
-
-    # final formatting of dask dataframes
-    if backend in ["dask", "dask_lazy"]:
-        import dask.dataframe as dd
-
-        metadata = _get_column_order_and_datatype(scoring, return_data)
-
-        results = dd.from_delayed(results, meta=metadata)
-        if backend == "dask":
-            results = results.compute()
-    else:
-        results = pd.concat(results)
-
-    # final formatting of results DataFrame
-    results = results.reset_index(drop=True)
-
-    return results
